feature(Match): 	1. 매칭요청서 validation check 및 카카오페이 결제 후 매칭 저장 구현
issue: #229

+++ backend/src/main/java/com/patriot/fourlipsclover/restaurant/entity/City.java
@@ -15,7 +15,7 @@ import lombok.Setter;
 public class City {
     @Id
     @Column(name = "city_id")
-    private int cityId;
+    private Long cityId;
 
     @Column(name = "name")
     private String name;
